[release] Rewrite index extraction logic in rust * Adds walkdir to the dependencies * There is also some moving around of files for the release module
