+++ Cargo.toml
@@ -39,6 +39,7 @@ ureq = { version = "3.1.4", default-features = false, features = ["rustls"], opt
 rkyv = { version = "0.8.15" }
 rayon = { version = "1.11.0", default-features = false }
 rusqlite = { version = "0.38.0", default-features = false, features = ["bundled"] }
+walkdir = "2.5.0"
 
 [dev-dependencies]
 criterion = "0.8.1"
