+++ src/dict/release/mod.rs
@@ -1,0 +1,5 @@
+mod core;
+mod db;
+mod index;
+
+pub use core::*;
